+++ app/Http/Controllers/Auth/LoginController.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace App\Http\Controllers\Auth;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+use Illuminate\Support\Facades\Auth;
+
+class LoginController extends Controller
+{
+    protected $redirectTo = '/';
+
+    public function __construct() {
+        $this->middleware('guest')->except('logout');
+    }
+
+    public function showLogin() {
+        return view('login');
+    }
+
+    public function login(Request $request) {
+        $credentials = $request->validate([
+            'email' => 'required|email',
+            'password' => 'required',
+        ]);
+
+        if (Auth::attempt($credentials)) {
+            $request->session()->regenerate();
+
+            return redirect()->intended('profile');
+        }
+
+        return back();
+    }
+
+    public function logout() {
+        Auth::logout();
+
+        return redirect('/');
+    }
+}
