+++ app/Http/Controllers/Auth/RegisterController.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace App\Http\Controllers\Auth;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
+
+use App\Models\User;
+
+class RegisterController extends Controller
+{
+    protected $redirectTo = '/';
+
+    public function __construct() {
+        $this->middleware('guest');
+    }
+
+    public function showRegister() {
+        return view('register');
+    }
+
+    public function register(Request $request) {
+        $request->validate([
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users',
+            'password' => 'required|string|min:6',
+        ]);
+
+
+        $user = User::query()->create([
+            'name' => $request['name'],
+            'email' => $request['email'],
+            'password' => Hash::make($request['password']),
+        ]);
+
+        Auth::login($user);
+
+        return redirect()->route('profile');
+    }
+}
