+++ app/Http/Controllers/UserController.php
@@ -10,53 +10,7 @@ use App\Models\User;
 
 class UserController extends Controller
 {
-    public function showRegister() {
-        return view('register');
-    }
-
-    public function register(Request $request) {
-        $user = User::query()->create([
-            'name' => $request['name'],
-            'email' => $request['email'],
-            'password' => Hash::make($request['password']),
-        ]);
-
-        Auth::login($user);
-
-        return redirect()->route('profile');
-    }
-
     public function profile() {
         return view('profile');
-    }
-
-    public function logout() {
-        Auth::logout();
-
-        return redirect('/');
-    }
-
-    public function showLogin() {
-        return view('login');
-    }
-
-    public function login(Request $request) {
-        $credentials = $request->validate([
-            'email' => [
-                'required',
-                'email',
-            ],
-            'password' => [
-                'required',
-            ],
-        ]);
-
-        if (Auth::attempt($credentials)) {
-            $request->session()->regenerate();
-
-            return redirect()->intended('profile');
-        }
-
-        return back();
     }
 }
